requirements.txt
envyaml
SQLAlchemy
psycopg2
llama-index-core
llama-index-embeddings-huggingface
llama-index-embeddings-openai
llama-index-llms-openai
streamlit
pandas